Extract SdkAvailable function to helper
